[tmux] #F flag only on current window
This is mostly so that I don't miss that some windows are zoomed. It
would still be nice that the 'Z' character (or any other visual cue) was
there rather than using it like this, because apart from the zoom option
I am not interested in virtually any of all the other options the #F
flag provides.