Trying fish shell after all
Would be really nice to not have to change shell for proper
autosuggestions, but the plugin for zsh for autosuggestions (by tarruda,
btw) is not really working for me. It's breaking other zsh stuff and
doesn't seem to be that stable and much supported. So, for those
reasons, going to try out the fish shell.